+++ treesapp/taxonomic_hierarchy.py
@@ -205,6 +205,18 @@ class TaxonomicHierarchy:
             logging.debug("Taxon name '{}' not present in taxonomic hierarchy.\n".format(prefix_taxon))
             return
 
+    def get_taxon_descendents(self, prefix_taxon: str) -> list:
+        """Returns a list of all Taxon instances with prefix_taxon in their lineage."""
+        descendents = []
+        ref_taxon = self.get_taxon(prefix_taxon)
+        if not ref_taxon:
+            logging.warning("No descendents of '{}' found as it's not in the hierarchy.\n".format(prefix_taxon))
+            return descendents
+        for hierarchy_taxon in self.hierarchy.values():  # type: Taxon
+            if ref_taxon in hierarchy_taxon.lineage():
+                descendents.append(hierarchy_taxon)
+        return descendents
+
     def get_bare_taxon(self, taxon_name: str):
         """
         A function for querying the hierarchy with a taxon that lacks a rank-prefix.
@@ -736,6 +748,9 @@ class TaxonomicHierarchy:
             if not ref_leaf.lineage:
                 continue
             taxa = ref_leaf.lineage.split(self.lin_sep)
+            if taxa[0] != self.root_taxon:
+                ref_leaf.lineage = self.root_taxon + self.lin_sep + ref_leaf.lineage
+                taxa = ref_leaf.lineage.split(self.lin_sep)
             while taxa:
                 taxon_name = taxa.pop(0)  # type: str
                 rank_prefix = taxon_name.split(self.taxon_sep)[0]
@@ -890,7 +905,7 @@ class TaxonomicHierarchy:
             lineage_split = lineage_split[:-1]
         return ref_lineage
 
-    def project_lineage(self, lineage_str):
+    def project_lineage(self, lineage_str) -> bool:
         """
         Function for checking whether all taxa in a lineage exist within TaxonomicHierarchy